// src/types.ts
export interface CustomOptionsType extends CustomReportOptions {
    jsErrors?: boolean
    apiErrors?: boolean
    resourceErrors?: boolean
    autoTracePerf?: boolean
    useFmp?: boolean
    enableSPA?: boolean
    vue?: any
    traceSDKInternal?: boolean
    detailMode?: boolean
    environment?: string
    perfInterval?: number
    authorization?: string
    teamId?: string
    enableDistributedTracing: boolean
    distributedTracingSkipUrls?: string[]
    enableDirectFetchPatching?: boolean
}

export interface UserContext {
    userId?: string
    countryCode?: string
}

export interface CustomReportOptions {
    collector?: string
    service: string
    pagePath: string
    serviceVersion: string
    countryCodes?: string[]
    authorization?: string
    teamId?: string
}